requirements.txt
django==3.0.4
pillow
djangorestframework
